Refactor Payments settings API controller business logic into separate service
