+++ plugins/woocommerce/src/Internal/DependencyManagement/ServiceProviders/AdminSettingsServiceProvider.php
@@ -3,8 +3,8 @@ declare( strict_types=1 );
 
 namespace Automattic\WooCommerce\Internal\DependencyManagement\ServiceProviders;
 
+use Automattic\WooCommerce\Internal\Admin\Settings\Payments;
 use Automattic\WooCommerce\Internal\Admin\Settings\PaymentsRestController;
-use Automattic\WooCommerce\Internal\Admin\Suggestions\PaymentExtensionSuggestions;
 
 /**
  * Service provider for the admin settings controller classes in the Automattic\WooCommerce\Internal\Admin\Settings namespace.
@@ -17,6 +17,7 @@ class AdminSettingsServiceProvider extends AbstractInterfaceServiceProvider {
 	 */
 	protected $provides = array(
 		PaymentsRestController::class,
+		Payments::class,
 	);
 
 	/**
@@ -25,7 +26,8 @@ class AdminSettingsServiceProvider extends AbstractInterfaceServiceProvider {
 	 * @return void
 	 */
 	public function register() {
+		$this->share( Payments::class );
 		$this->share_with_implements_tags( PaymentsRestController::class )
-			->addArgument( PaymentExtensionSuggestions::class );
+			->addArgument( Payments::class );
 	}
 }
